backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "cryptography>=46.0.5",
    "fastapi>=0.128.8",
    "openai>=2.29.0",
    "pydantic-settings>=2.11.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.39.0",
    "websockets>=15.0.1",
]
